refactor: useTags 구현
